refactor: move trim, toLower from extensions
