+++ src/IFingerprintBuilder.cs
@@ -7,6 +7,8 @@ public interface IFingerprintBuilder<T>
 {
     IFingerprintBuilder<T> For<TProperty>(Expression<Func<T, TProperty>> expression);
 
+    IFingerprintBuilder<T> For(Expression<Func<T, string>> expression, bool toLower, bool trim);
+
     IFingerprintBuilder<T> For<TProperty>(Expression<Func<T, TProperty>> expression, Expression<Func<TProperty, string>> fingerprint);
 
     Func<T, byte[]> Build();
